requirements.txt
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary
django-redis
parameterized==0.9.0
pillow==11.3.0
PyJWT==2.9.0
sqlparse==0.5.3
timedelta==2020.12.3
tzdata==2025.2
mysqlclient
dotenv==0.9.9
python-dotenv==1.1.1
